fix(api-key): address Copilot review on #270
- list: scope ranking-overview key fetch to the current workspace via resource
  meta and drop the 500-row page cap (pagination mode off) so the donut totals
  and top-3 attribution stay correct and don't over-fetch across workspaces.
- limits card: show backend-computed token_quota.remaining (source of truth,
  can convey overage) instead of recomputing limit - used.
- limitsToForm: validate token_quota.period against the known periods before
  using it as the combobox value; fall back to the default otherwise.
- types: define ApiKeyLimits in domain types.ts (and import it from there in the
  hook/components) to remove the types -> hooks layering dependency.
