+++ .i18n-tracker.lock
@@ -39,7 +39,7 @@
   "src/pages/clusters/edit.tsx": "ecb3a6e4f04286c161d3872d9e91bfa9",
   "src/pages/clusters/list.tsx": "fd9a37d70069e11ba24032afd2d735c0",
   "src/pages/clusters/show.tsx": "4cf0a385edd011ab6b722d4db4f88de2",
-  "src/pages/api-keys/list.tsx": "bb999a0ffb607b443d513771635653a0",
+  "src/pages/api-keys/list.tsx": "63ef2b7cd9931d1241c23f6c51a7eeef",
   "src/pages/api-keys/show.tsx": "3ecbea3750b6bc421c97fb673084dd5f",
   "src/components/ui/alert-dialog.tsx": "fd2a0854b18d61d21fe0e8effd7e4cd9",
   "src/components/ui/alert.tsx": "25ac92016d099c561981ce1456ee8741",
@@ -203,7 +203,7 @@
   "src/domains/cluster/components/ClusterStatus.tsx": "532287ae3d6040ed5dbfc7beac04f524",
   "src/domains/cluster/components/ClusterType.tsx": "439521dd989dcf3b6054bdea68e542e8",
   "src/domains/cluster/components/NodeIPsField.tsx": "26458bb02e6907bcba9e743b5d9c52ca",
-  "src/domains/api-key/types.ts": "ae6c0f6ed0366fcabe77ff04df32fc07",
+  "src/domains/api-key/types.ts": "1e1e02893bf31e7ac9f2c7cfdd6274ea",
   "src/foundation/lib/yaml-transform.ts": "dc8dc9b4f12c8a4565cb2ad4b21e286f",
   "src/foundation/lib/grafana-dashboard-url.ts": "a457d0bef7f683a9185b163f79630278",
   "src/foundation/lib/sidebar-active.ts": "d4a316c595e9fc8d751d6b7961d26023",
@@ -309,10 +309,10 @@
   "src/foundation/components/GpuDeviceResourcesView.tsx": "59fef5bb066810a55e6c7e82f2f9c1ac",
   "src/domains/endpoint/components/EndpointClusterGpuResourcesPanel.tsx": "92c8ad14dcce853cb91732c0773196f3",
   "src/domains/cluster/lib/accelerator-virtualization.ts": "24f5619a2206dfc0b1da6757428fb149",
-  "src/domains/api-key/components/ApiKeyLimitsCard.tsx": "e6b761d99dfbfff132a6fd6ef5f19896",
+  "src/domains/api-key/components/ApiKeyLimitsCard.tsx": "6ff7c0b75f0dd7a19b8583716fab5b5a",
   "src/domains/api-key/components/ApiKeyPerformanceCard.tsx": "986c799c7c0ca0993eb7c99ae3880a12",
   "src/domains/api-key/components/ApiKeyPolicyFields.tsx": "a497e28db1b73d61bb1e4991b507eac4",
   "src/domains/api-key/components/ApiKeyRankingOverview.tsx": "6f7cfca5827d555328cd2ca35585c521",
   "src/domains/api-key/components/ModelMultiSelect.tsx": "7cc524720fdcadde0d72f3cd49afe1bb",
-  "src/domains/api-key/hooks/use-api-key-policy.ts": "4334710d359f1c6df7cefb71b14863d6"
+  "src/domains/api-key/hooks/use-api-key-policy.ts": "01a3df4ad6c50e3572bd74d56cb8f581"
 }
